# docs/openapi.json
{"openapi": "3.1.0", "info": {"title": "VectorAPI", "description": "VectorAPI is a service for managing vector collections and performing vector similarity queries using a PostgreSQL vector database with the `pgvector` extension. Utilizes `fastapi` for the HTTP API, `pgvector` and SQLAlchemy for the vector database side and relies on `pytorch` for computing embeddings.\n\nhttps://github.com/grafana/vectorapi\n", "version": "0.1.0"}, "paths": {"/v1/embeddings": {"post": {"tags": ["embeddings"], "summary": "Create Embeddings", "description": "Create embeddings for a given text.", "operationId": "create_embeddings_v1_embeddings_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmbeddingRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmbeddingResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/similarity": {"post": {"tags": ["embeddings"], "summary": "Similarity", "description": "Calculate similarity between two inputs", "operationId": "similarity_v1_similarity_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SimilarityRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"type": "number"}, "type": "array", "title": "Response Similarity V1 Similarity Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/collections/create": {"post": {"tags": ["collections"], "summary": "Create Collection", "description": "Create a new collection with the given name and dimension.", "operationId": "create_collection_v1_collections_create_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateCollectionRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/collections/delete": {"post": {"tags": ["collections"], "summary": "Delete Collection", "description": "Delete a collection with the given name.", "operationId": "delete_collection_v1_collections_delete_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeleteCollectionRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/collections/{collection_name}": {"get": {"tags": ["collections"], "summary": "Get Collection", "description": "Get a collection with the given name.", "operationId": "get_collection_v1_collections__collection_name__get", "parameters": [{"name": "collection_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Collection Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/collections": {"get": {"tags": ["collections"], "summary": "List Collections", "description": "List all collections.", "operationId": "list_collections_v1_collections_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/v1/collections/{collection_name}/upsert": {"post": {"tags": ["points"], "summary": "Upsert Point", "description": "Create a new collection with the given name and dimension.", "operationId": "upsert_point_v1_collections__collection_name__upsert_post", "parameters": [{"name": "collection_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Collection Name"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CollectionPointRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/collections/{collection_name}/delete/{point_id}": {"delete": {"tags": ["points"], "summary": "Delete Point", "description": "Delete a collection point with the given id.", "operationId": "delete_point_v1_collections__collection_name__delete__point_id__delete", "parameters": [{"name": "collection_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Collection Name"}}, {"name": "point_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Point Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/collections/{collection_name}/get/{point_id}": {"get": {"tags": ["points"], "summary": "Get Point", "description": "Get the collection point matching the given id.", "operationId": "get_point_v1_collections__collection_name__get__point_id__get", "parameters": [{"name": "collection_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Collection Name"}}, {"name": "point_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Point Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/collections/{collection_name}/query": {"post": {"tags": ["points"], "summary": "Query Points", "description": "Query collection with a given embedding query.", "operationId": "query_points_v1_collections__collection_name__query_post", "parameters": [{"name": "collection_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Collection Name"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QueryPointRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/collections/{collection_name}/search": {"post": {"tags": ["points"], "summary": "Search", "description": "Search collection with a given text input.", "operationId": "search_v1_collections__collection_name__search_post", "parameters": [{"name": "collection_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Collection Name"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SearchPointRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"CollectionPointRequest": {"properties": {"id": {"type": "string", "title": "Id"}, "input": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Input"}, "embedding": {"anyOf": [{"items": {"type": "number"}, "type": "array"}, {"type": "null"}], "title": "Embedding"}, "metadata": {"type": "object", "title": "Metadata", "default": {}}, "model": {"type": "string", "title": "Model", "default": "BAAI/bge-small-en-v1.5"}}, "type": "object", "required": ["id"], "title": "CollectionPointRequest"}, "CreateCollectionRequest": {"properties": {"collection_name": {"type": "string", "title": "Collection Name"}, "dimension": {"type": "integer", "title": "Dimension"}, "exist_ok": {"type": "boolean", "title": "Exist Ok", "default": false}}, "type": "object", "required": ["collection_name", "dimension"], "title": "CreateCollectionRequest"}, "DeleteCollectionRequest": {"properties": {"collection_name": {"type": "string", "title": "Collection Name"}}, "type": "object", "required": ["collection_name"], "title": "DeleteCollectionRequest"}, "EmbeddingRequest": {"properties": {"model": {"type": "string", "title": "Model", "default": "BAAI/bge-small-en-v1.5"}, "input": {"type": "string", "title": "Input"}, "user": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "User"}}, "type": "object", "required": ["input"], "title": "EmbeddingRequest"}, "EmbeddingResponse": {"properties": {"object": {"type": "string", "title": "Object", "default": "list"}, "data": {"items": {"$ref": "#/components/schemas/EmbeddingResponseData"}, "type": "array", "title": "Data"}, "model": {"type": "string", "title": "Model"}, "usage": {"$ref": "#/components/schemas/EmbeddingResponseUsage"}}, "type": "object", "required": ["data", "model", "usage"], "title": "EmbeddingResponse"}, "EmbeddingResponseData": {"properties": {"index": {"type": "integer", "title": "Index"}, "object": {"type": "string", "title": "Object", "default": "embedding"}, "embedding": {"items": {"type": "number"}, "type": "array", "title": "Embedding"}}, "type": "object", "required": ["index", "embedding"], "title": "EmbeddingResponseData"}, "EmbeddingResponseUsage": {"properties": {"promptTokens": {"type": "integer", "title": "Prompttokens"}, "totalTokens": {"type": "integer", "title": "Totaltokens"}}, "type": "object", "required": ["promptTokens", "totalTokens"], "title": "EmbeddingResponseUsage"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "QueryPointRequest": {"properties": {"query": {"items": {"type": "number"}, "type": "array", "title": "Query"}, "top_k": {"type": "integer", "title": "Top K", "default": 10}, "filter": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Filter"}}, "type": "object", "required": ["query"], "title": "QueryPointRequest"}, "SearchPointRequest": {"properties": {"input": {"type": "string", "title": "Input"}, "filter": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Filter"}, "top_k": {"type": "integer", "title": "Top K", "default": 10}, "model": {"type": "string", "title": "Model", "default": "BAAI/bge-small-en-v1.5"}}, "type": "object", "required": ["input"], "title": "SearchPointRequest"}, "SimilarityRequest": {"properties": {"model": {"type": "string", "title": "Model", "default": "BAAI/bge-small-en-v1.5"}, "sourceSentence": {"type": "string", "title": "Sourcesentence"}, "sentences": {"items": {"type": "string"}, "type": "array", "title": "Sentences"}}, "type": "object", "required": ["sourceSentence", "sentences"], "title": "SimilarityRequest"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}, "tags": [{"name": "embeddings", "description": ""}, {"name": "collections", "description": "A collection represents a set of vectors with metadata, which can be queried and searched. In postgres, they're are stored as tables under the schema defined by the `VECTORAPI_STORE_SCHEMA` environment variable (default: `vectorapi`).\n"}, {"name": "points", "description": "CollectionPoints are the individual vectors (and their metadata) that make up a collection.\n\n### Metadata search filters\n\nThe API supports filtering the search results by metadata, through the `filter` parameter. This can be set in both the `/query` (search by vector) and `/search` (search by text) endpoints.\n\nSupported single value filter operators are:\n\n- Equality: `$eq`\n- Inequality: `$ne`\n\nSupported multiple values filter operators are:\n\n- `$or`: The `$or` operator supports passing multiple filters to the query call\n- `$and`\n\n#### Example vector search request with simple match filter\n\nEndpoint: POST http://localhost:8889/v1/collections/{collection_name}/search\n\n```json\n{\n  \"input\": \"A small phone with a great camera\",\n  \"filter\": {\n    \"category\": {\n      \"$eq\": \"electronics\"\n    }\n  }\n}\n```\n\n#### Example vector search request with OR filter\n\nEndpoint: POST http://localhost:8889/v1/collections/{collection_name}/query\n\n```json\n{\n  \"query\": [0.4, 0.5, 0.6],\n  \"filter\": {\n    \"$or\": [\n      {\n        \"category\": {\n          \"$eq\": \"electronics\"\n        }\n      },\n      {\n        \"category\": {\n          \"$eq\": \"clothing\"\n        }\n      }\n    ]\n  }\n}\n```\n"}]}